Refactor Body component to utilize custom hooks for restaurant data fetching and searching; remove redundant state management
